Find command: find name and tags
Find command only checks name for keywords.

A find which can search both names and tags is more user-friendly as
users can search for people through their names or what they are
tagged as.

Let's update the search algorithm to search both names and tags and
rename the predicate handling the search to better reflect what it
now does.