Fix alias bug
A bug was reported that prefix of existing command should show error
message, despite of the warning given in the user guide. Another bug
reported is that alias can contain existing command word.

Alias replacement is done if it matches a word instead of substring.
Additional check is created when parsing addalias command.

It helps to create a more robust system.

Resolves: #151, #142